@@ -8,20 +8,16 @@ Gem::Specification.new do |s|
 
   s.files       = `git ls-files -z`.split("\x0")
   s.executables = s.files.grep(%r{^bin/}) { |f| File.basename(f) }
-  s.test_files  = s.files.grep(%r{^(tests|spec)/})
 
   s.homepage    = 'https://github.com/voxpupuli/metadata-json-lint'
   s.license     = 'Apache-2.0'
 
-  s.required_ruby_version = '>= 2.5.0'
+  s.required_ruby_version = '>= 2.7.0'
 
   s.add_runtime_dependency 'json-schema', '>= 2.8', '< 5.0'
   s.add_runtime_dependency 'semantic_puppet', '~> 1.0'
   s.add_runtime_dependency 'spdx-licenses', '~> 1.0'
   s.add_development_dependency 'rake', '~> 13.0', '>= 13.0.6'
   s.add_development_dependency 'rspec', '~> 3.12'
-  s.add_development_dependency 'rubocop', '~> 1.28.0'
-  s.add_development_dependency 'rubocop-performance', '~> 1.10'
-  s.add_development_dependency 'rubocop-rake', '~> 0.2'
-  s.add_development_dependency 'rubocop-rspec', '~> 2.10.0'
+  s.add_development_dependency 'voxpupuli-rubocop', '~> 1.2'
 end
