@@ -18,7 +18,7 @@ else
 
   SimpleCov.formatters = [
     SimpleCov::Formatter::Console,
-    SimpleCov::Formatter::Codecov
+    SimpleCov::Formatter::Codecov,
   ]
 end
 
